+++ produits_derives_lidar/ip_one_tile.py
@@ -8,11 +8,12 @@ import os
 import tempfile
 
 import hydra
-from omegaconf import DictConfig, OmegaConf
-from pdaltools.las_info import parse_filename
+from omegaconf import DictConfig
 
 from produits_derives_lidar.commons import commons
-from produits_derives_lidar.tasks.las_interpolation_deterministic import Interpolator
+from produits_derives_lidar.tasks.las_interpolation_deterministic import (
+    interpolate_from_config,
+)
 from produits_derives_lidar.tasks.las_raster_generation import (
     mask_with_no_data_shapefile,
 )
@@ -20,65 +21,9 @@ from produits_derives_lidar.tasks.las_raster_generation import (
 log = commons.get_logger(__name__)
 
 
-@commons.eval_time_with_pid
-def interpolate(input_file: str, output_raster: str, config: dict):
-    """Run interpolation
-    Args:
-        input_file(str): File on which to run the interpolation
-        output_raster(str): output file for raster image (with buffer)
-        config(dict): dictionary that must contain
-                {
-                    "tile_geometry": {
-                        "tile_coord_scale": #int,
-                        "tile_width": #int,
-                        "pixel_size": #float,
-                        "no_data_value": #int
-                    },
-                    "io": {
-                        "spatial_reference": #str
-                    },
-                    "interpolation": {
-                        "algo_name": #str
-                    }
-                }
-            with
-                tile_coord_scale value(int): coords in tile names are in km
-                tile_width value(int): tile width in meters
-                pixel_size value(float): pixel size for raster generation
-                spatial_ref value(str): spatial reference to use when reading las file
-                interpolation_method value(str): interpolation method for raster generation
-    Output:
-        ras: output raster (WARNING: can be None for some methods)
-        origin: tile origin
-        can_interpolate (bool): false if there were no points to interpolate
-    """
-    _, coordX, coordY, _ = parse_filename(input_file)
-    origin = [
-        float(coordX) * config["tile_geometry"]["tile_coord_scale"],
-        float(coordY) * config["tile_geometry"]["tile_coord_scale"],
-    ]
-    nb_pixels = [
-        int(config["tile_geometry"]["tile_width"] / config["tile_geometry"]["pixel_size"]),
-        int(config["tile_geometry"]["tile_width"] / config["tile_geometry"]["pixel_size"]),
-    ]
-
-    _interpolation = Interpolator(
-        nb_pixels,
-        origin,
-        config["tile_geometry"]["pixel_size"],
-        config["interpolation"]["algo_name"],
-        config["io"]["spatial_reference"],
-        config["tile_geometry"]["no_data_value"],
-        config["tile_geometry"]["tile_width"],
-        config["tile_geometry"]["tile_coord_scale"],
-        config["filter"]["keep_classes"],
-    )
-    _interpolation.run(input_file, output_raster)
-
-
 @hydra.main(config_path="../configs/", config_name="config.yaml", version_base="1.2")
 def run_ip_on_tile(config: DictConfig):
-    """Run interpolation on single tile useing hydra config
+    """Run interpolation on single tile using hydra config
     config parameters are explained in the default.yaml files
     """
     if config.io.input_dir is None:
@@ -97,24 +42,20 @@ def run_ip_on_tile(config: DictConfig):
 
     # for export
     _size = commons.give_name_resolution_raster(config.tile_geometry.pixel_size)
-    geotiff_stem = f"{tilename}{_size}_{config.interpolation.method_postfix}"
+    geotiff_stem = f"{tilename}{_size}"
     geotiff_filename = f"{geotiff_stem}.tif"
     geotiff_path = os.path.join(config.io.output_dir, geotiff_filename)
-
-    # process
-    dico_config = OmegaConf.to_container(config)
-    interpolate(input_file, geotiff_path, dico_config)
 
     if config.io.no_data_mask_shapefile:
         with tempfile.NamedTemporaryFile(suffix=".tif", prefix=f"{geotiff_stem}_raw") as tmp_geotiff:
             # process interpolation
-            interpolate(input_file, tmp_geotiff.name, config)
+            interpolate_from_config(input_file, tmp_geotiff.name, config)
             mask_with_no_data_shapefile(
                 config.io.no_data_mask_shapefile, tmp_geotiff.name, geotiff_path, config.tile_geometry.no_data_value
             )
 
     else:
-        interpolate(input_file, geotiff_path, config)
+        interpolate_from_config(input_file, geotiff_path, config)
 
 
 def main():